app_v4_new/requirements.txt

numpy
librosa
soundfile
pyloudnorm; python_version >= "3.8"
python-dotenv
requests
yt-dlp
mysql-connector-python
